test: ensure that task listener job is cleaned up
When terminating the element instance, the listener job must be canceled
in order for it to be removed from the state, so it is no longer
activatable.

Such tests already exist for execution listener (START and END).

I see no need to explicitly test this for assigning and completing
separately, but feel free to challenge this.