test: ensure incident resolved on termination
This adds a test case for both task listeners as well as execution
listeners to ensure that incidents are cleaned up when the associated
element instance is terminated.

This is achieved using the Incident.RESOLVED event.

Note that this functionality already exists for execution listeners, but
I could find no test case for it.

For task listeners, this test currently fails.