fix: start calledProcessDepth at 0
There's a reason to keep it at -1, if you want to differentiate between
the calledProcessDepth not set, and it just being set by default.

The -1 case could've been useful to ensure we set the calledProcessDepth
correctly for child process instance created after 8.7 that are part of
root process instance created prior to 8.7. However, this comes at the
cost of performance, as a lookup is needed for each called level.

Instead, we simply set the default to 0, to ensure that all element
instances created prior to 8.7, have it set to 0.

This has the added benefit that we do not have to explicitly set the
value in other places (like test code).

Co-authored-by: Dmitriy Melnychuk <dmitriy.melnychuk@capgemini.com>
Co-authored-by: Remco <remco@westerhoud.nl>
Co-authored-by: Stephan Epping <stephan.epping@camunda.com>
Co-authored-by: berkaycanbc <berkay.can@camunda.com>