feat: exchange heartbeat information on channel creation to avoid false-positives
HeartbeatSetupHandler.Client is responsible for sending heartbeat
informations to the server.
HeartbeatSetupHandler.Server receives these informations and replies to
the client if heartbeats are enabled or not.

The client, upon receiving when receiving heartbeatsEnable=true adds to
the ChannelPipeline the HeartbeatHandler.Client, which registers
autonomusly the IdleStateHandler.

The server will add to the ChannelPipeline the HeartbeatHandler.Server
when it receives the first HeartbeatSetupRequest.

Both HeartbeatSetupHandler remove themselves from the ChannelPipeline
when the setup has been done.

The HeartbeatSetup{Request,Response} are encoded using a new schema
defined in "cluster-messaging.xml"

closes #26879