fix: switch to process depth
The called process depth could be understood at the depth of called
processes, implying that it starts at 0 and increments with every called
process.

This makes the config option a bit confusing, because a max depth of 10
is more easily understood as 10 levels of processes rather than 10
called processes (11 levels of processes).

To avoid this, we've changed the property to be named `processDepth` and
increased the default value to 1.

Co-authored-by: Dmitriy Melnychuk <dmitriy.melnychuk@capgemini.com>
Co-authored-by: Remco <remco@westerhoud.nl>
Co-authored-by: Stephan Epping <stephan.epping@camunda.com>
Co-authored-by: berkaycanbc <berkay.can@camunda.com>
(cherry picked from commit e969f7f66fcb443d15e10c4052045c6f060cb355)