test: ensure incident is raised at max depth
We don't want to create a child instance exceeding the maximum depth.
So, we should be raising the incident at the call activity that exists
at the maximum depth itself.

We can test this by looking at the depth variable which is a local
variable created at the call activity where the value is the current
depth + 1 (it is propagated as the depth of the child instance).

Co-authored-by: Dmitriy Melnychuk <dmitriy.melnychuk@capgemini.com>
Co-authored-by: Remco <remco@westerhoud.nl>
Co-authored-by: Stephan Epping <stephan.epping@camunda.com>
Co-authored-by: berkaycanbc <berkay.can@camunda.com>