refactor: clean up test
processInstanceKey was unused.

incidentRecords were not limited to the specific incident we're
interested in. Typically we limit records to the specific process
instance, but that's a bit difficult here (due to the called instances).
Instead, we can simply limit on the element id as it's not used in the
other test cases.

Aligned the text block.

Co-authored-by: Dmitriy Melnychuk <dmitriy.melnychuk@capgemini.com>
Co-authored-by: Remco <remco@westerhoud.nl>
Co-authored-by: Stephan Epping <stephan.epping@camunda.com>
Co-authored-by: berkaycanbc <berkay.can@camunda.com>
(cherry picked from commit f89724034f733506e4829517144337210b13b24b)