refactor: clarify the max depth comparison
This compares the next (called) process' depth with the max depth. So it
makes more sense to compare the current + 1 with the allowed max,
instead of comparing the current with the allowed max but allowing them
to be equivalent.

(cherry picked from commit d839ad66950ea40ce9fcacfce598e91d54edde96)