refactor: split ElementActivatingApplier versions
Splitting the ElementActivatingApplier is necessary to ensure that the
process depth is only set for child instances created on 8.7. If we
don't split, we have a chance that child instance created prior to 8.7,
still get a different depth set because they are replayed differently on
8.7.

This is simply renaming ProcessInstanceElementActivatingApplier to
include a version number, and copy it to another class with a different
version number. Both classes are identical apart from the name.

In the next commit, we can revert V1 back to its state before our
changes.

Normally, I'd do this the other way around to make the review easier,
but that wasn't possible anymore without many conflicts. I still wanted
to find a way to highlight the changes we've made.