docs: clarify subscription timeout property
This property is used as the sent time which is used again by the
visit implementation. I felt it was confusing so I added some
documentation to clarify what this is.

I also figured out that the TODO was not relevant anymore.

(cherry picked from commit 34a168e254f79368d50c3a4b473604009525a7e5)