fix: only correlate message to subscription once
The message key is already stored in a process message subscription,
when it is correlated. So we can simply compare it to the message key
that we're currently trying to correlate.

This does not fully resolve the bug yet, because two messages may
correlate to the same subscription (as it is non-interrupting) and no
ordering for arrival at this partition is guaranteed for their CORRELATE
commands.

However, this does already resolve the bug in case when the correlations
arrive in order naturally. I.e., when the message is published, and
correlates, and then some time later, another is published that
correlates to this subscription.

It does not resolve the case, where multiple messages are buffered when
the subscription opens.