test: ensure message correlation is idempotent
Message correlations are retried if not acknowledged. We must ensure
that CORRELATE commands that are retried one or multiple times, still
just trigger the ProcessMessageSubscription once, even if it concerns a
non-interrupting catch event.

This test case publishes 10 messages that each should correlate to a
non-interrupting message event subprocess. However, the partition with
the process instance is paused, causing the message partition to retry
sending the correlate commands.

The test verifies after processing is started again, the catch event is
triggered exactly 10 times.

At this time, the test fails. We'll work to fix the related bug in
subsequent commits.

Co-authored-by: Remco <remco@westerhoud.nl>