feat: before taking a Backup check if all required indices are present
Before taking a backup check in ES/OS which indices are present. If a
required index is missing, return an error.
If a non-required index is missing, just skip it. If there is no index
for a particular part, skip that part altogether.
Skipping empty parts avoids doing a snapshot of the entire database

relates #26618