test: In AuthorizationCheckBehaviourTest run actions on the same thread as the engine
Before this change, authorizationCheckBehaviour interact with mutables
classes that are used concurrently by the engine in another thread.

All methods called in authorizationCheckBehaviour are run in the same
thread as the StreamProcessor.
Although this may not be the case of the SIGSEV, it can help us identify
if it's generated from somewhere else if it's not fixed by this commit.