test: verify controllers always call withAuthentication when accessing a service
This ArchUnit test verifies that every call to a Service that is performed by a Controller adds the authentication by
calling withAuthentication first. The test will fail if this is not the case.

This is essential, as not making this call could result in cases where forget to add it, or accidentally remove it (for
example during a merge conflict). As a result the endpoint would not verify a user's authorization anymore. Potentially
causing data-leaks.