fix: store error catch event trigger correctly
Store it in the process event key. Use it to store a newly generated key
that can be used later for the boundary event instance.

Use the process event key to look up the key of the boundary event
instance to create.

Finally look up the catch event using the scope key.