fix: reject unnecessary correlate acknowledgement
Rather than dealing with an unnecessary acknowledgement in the
CORRELATED event applier, we can simply avoid appending the CORRELATED
event under the conditions where it was already CORRELATED previously.

The conditions under which this is the case doesn't change, so it's
still if the message key of the acknowledgement doesn't match the
message key currently on the message subscription, or if the message
subscription is already not correlating anymore.

This has the added benefits that:
- we don't have to introduce a new version of the event applier, as we
  don't have to make change to it to fix the bug
- we can clarify the behavior on the log stream by appending a command
  rejection instead of the CORRELATED event.

This ensures the bug is resolved for newly processed CORRELATE commands.

(cherry picked from commit 15309eaf8230f6e17764ea7e3477b3bf91f25d46)