docs: remove no longer needed todo
We can now ensure that the correlations arrive in order, because only a
single message can be correlating to a subscription. If a second message
arrives that wants to correlate to the same subscription, it has to wait
on the message partition until the first message has been fully
correlated.