test: simplify test case
There's no need to manually create an admin with all permissions. Zeebe can create this automatically given the right
config is provided.

With this commit I've added the admin user to the initialization configuration. As a result the the admin user gets
assigned to the admin role and automatically receives all the available permissions.