refactor: enforce explicit error type mapping in `IncidentEntityTransformer`
Removed the `default` case in the `toErrorType` switch expression to enforce a compile-time error:
> "the switch expression does not cover all possible input values"
instead of a runtime exception:
> "java.lang.IllegalArgumentException: Unexpected value: NEW_ERROR_TYPE"
when a new error type is added to the `io.camunda.webapps.schema.entities.operate.ErrorType` enum.

This ensures that developers must explicitly define the mapping for new values or explicitly throw an exception.

Additionally, updated the `describedAs` message for the recently added test by removing the note about updating
`IncidentEntityTransformer#toErrorType`, as the compiler will now clearly indicate where a new type should be mapped.