refactor: allow custom command writing
The CommandWriter interface available in our test clients typically
provides some overloaded methods to write a command or write a command
on a partition. However, as our needs for writing commands with more and
more specific record properties has grown, so have the number of these
methods.

In the MessageCorrelationMultiplePartitionsTest test case we need to
write a command to a specific partition, and so we've added a new create
method on the client that allows this. However, the client typically
provides a fluent builder pattern. I wanted to apply that here as well,
so we could write .onPartition(x).create() similar to usage in other
test clients (like PublishMessageClient). However, I quickly needed to
add new methods to the CommandWriter. In a future commit, we'll see that
I have even more methods to add the CommandWriter.

To avoid all this, I've added a generic writeCommandOnPartition method
that allows us to customize the record.

I would've liked to go a step further and remove the create(AuthInfo)
method as well, but perhaps that's for another time.

Note that I chose AuthInfo with claim AUTHORIZED_ANONYMOUS_USER for the
default case. This was required to pass tests that simply wanted to
create a process instance without authorization kicking in. This
happened because after this refactoring, the instance is created with
a request id. Previously this was dependent on the used method. However,
all process instance creation commands are expected to be part of a
request, so this solution seemed more fitting.

(cherry picked from commit 519e3741601e60db025b39eb013dfd242377d92f)