test: write complete user task to partition by key
In our test case, we need to write the complete user task command to
complete the process instance, which in turn allows us to easily limit
the recorded record stream. However, the instance with the user task
exists on partition 2, and our test client is unable to write the
command to partition 2. So, we'll need to add a new method to achieve
this.

We can use the new custom writeCommandOnPartition method to achieve this
while also passing the requestId, requestStreamId and the tenant ids as
authorizations, without having to add a new method to CommandWriter.

We know the key of the user task, so we know the partition id where we
want to write the command (can be decoded from the key).