test: specify Intents in test cases
We are generating records with a specific ValueType, but with a random Intent that belongs to the ValueType. There is
no guarantee the exporter handles events with this Intent. The Intent could even be a command instead of an event.

By explicitly specifying the Intent we can ensure the exporter will handle the record.