test: use message key on msg-sub correlate command
The MessageSubscription CORRELATE command now requires the key of the
message that was correlated to the subscription to ensure we can
deduplicate the acknowledgement. Without the key we could not.

These test cases started failing after we started rejecting the
CORRELATE acknowledgement when the message key differs.

(cherry picked from commit 7dc6fe704a05314b13f4778c920699496c4425a6)