test: test usage of misconfigured exec listener
This adds a hacky test case, that will fail due to a deployed
process with a misconfigured execution listener.

The restart of the broker (with higher version) fails because it's
unable to replay some events where the faulty process model is
transformed.

This is a similar case to that in SUPPORT-25748 and can be used to
attempt to patch the problem.