test: ignore misconfigured execution listeners
We've encountered a bug where misconfigured execution listeners could be
deployed. When transformed in later versions, this led to issues because
a transformer throwing an uncaught exception can ban an instance or even
kill a partition (on replay).

This adds some fast running unit tests that assert that misconfigured
execution listeners are ignored.