wip: split set variable document in two parts
To trigger a listener and not immediately set the variables (for a later
rollback on deny), we need to introduce new CREATING and UPDATING
variable events.

We also need to split the VariableDocumentUpdateProcessor to process two
commands: UPDATE and COMPLETE_TASK_LISTENER.

This also requires the JobCompleteProcessor, to somehow know that this
doesn't concern a regular user task listener, but a variable update
command. This is currently where I'm stuck. From a Complete Job record
you do not know what set of variable document changes it belonged to.
There's no reference to a set of variable changes.