wip: try reusing update user task flow
Rather than trying to turn Set Variable API into a User Task Listener
supported command, we can re-use the Update User Task command flow for
the Set Variable API. This means we're immediately re-using the task
listener structure as well.

We do need to ensure we can respond to the request with a fitting
response. This will require us to store some information about the
request and the specific intent early on, and use that information at
the end again.

We also need to store the variables on the `UPDATING` event of a user
task rather than discard them.

And finally we can merge the provided variables into the local variable
state when writing `UPDATED`.