fix: also consider message keys prior to the last
Since messages are ordered, we can consider a message to have been
correlated not just if the keys are equal, but also if the key is less
than the last correlated one.

Note that this case should not happen, because we expect message
correlations to have been completed before the next starts. However,
this additional case gives us some small safeguard against faults, and
should be easier to understand semantically:
- any message with a key equal to or less than the last correlated
  message key should be considered already correlated.