feat: create BasicAuthCredentialsProvider
This credentials provider will make it a better UX for people using basic auth. Instead of having to create the
authorization header themselves they only have to specify using this credentials provider and give it the username and
password. This provider will take care of encoding it properly and turning it into a valid authorization header.