test: verify builder validations and environment
Adds test cases to ensure the username and password are set and not empty.

This also adds testcases to verify the environment variable overrides work as expected. If the variables are set the
client will use these for the basic auth. This behaviour can be changed by setting the apply environment variables flag
to false.