feat: add stable/8.7 to testable branches
Due to this unique scenario where we have 2 development branches we need to test not only on main but also on the other
development branch (stable/8.7). This branch has been hardcoded in the workflow files. Once 8.7 is released we need to
remove those lines, to go back to the normal workflows