test: msg subscriptions are correctly created on multi-instance body migration
When the target process contains multi-instance body with event
subprocess, message subscription should be created for each inner
activity of the multi-instance body. With the existing bug, a message
start event subscription is attempted to be created for the
multi-instance body itself, which is against BPMN specification.
Added test case verifies that the message subscriptions are only
created for inner activities.