feat: set multi-instance body type explicitly during migration
The expected element type for multi-instance body needs to be set
explicitly because inner activities has the same element id.

If we do not set the expected element type, the inner activity of the
multi instance body is returned. As a result, event subscriptions are
handled for wrong target element.