fix: properly set the user key on user update
We set the user key to the key that was on the command. This will always be -1 since it's a command received from the
gateway. As users are identified by their username there's no need to have a userkey set.

We should make sure on followup commands that we set the userkey properly. We can do this by taking it from the
persisted user and setting it on the command. As well as ensuring that we write any followup commands/events with the
correct user key.