feat: make tenants CF id-based
The identifier for tenants got changed from keys to ids. The column family should reflect this.
As a result all the methods in the state that interact with this CF need to be changed as well. There were a bunch of
usages where the tenant got retrieved by key. These usages are switched to work with an id instead.