refactor: ensure stateful gauges are properly propagated
Had to revert the reflection usage which allowed us to push the
`StatefulGauge` custom meters down to Micrometer. The reason for this
is it completely breaks with the `CompositeMeterRegistry`. The composite
registry expects all its meters to implement a private interface which
is not accessible, so there's no way for us to hook into this lifecycle
except to use Java proxies - at that this point, it's just not worth it.

So instead, we add a new extension to the `CompositeMeterRegistry` which
can keep track of stateful gauges, and we keep the stateful gauge
interface for convenience: it will work with any meter registry, it's
just not kept identical.

By default, all wrapped registries in the brokers/gateway will be
stateful, so it should help mitigate issues around double gauge
registration.

(cherry picked from commit 4083607a917c58491f2af216d076bf4ec1a8b1b3)