docs: explain rebasing instead of merge commits
We advise rebasing rather than using a merge commit to update a branch,
but it was hidden inside the section where a merge was already queued.
That's not always the case; sometimes, folks need to update their branch
before it is approved.

Therefore, we should highlight this more explicitly. I've also explained
why this matters and improved the wording from 'rebase `main` into the
branch' to 'rebase onto the target branch'. Note that officially, the
target branch is called the 'base', but 'target' is probably a more
generally understood term.