fix: don't log job key twice
For JobBatch Record, we log the key of each job in the batch at the
start by providing the jobKey parameter.

However, for regular Job records, this means the record key is logged
and again at the start of the record value.

We can avoid this by passing -1 as parameter.

Note that we should align the spaces correctly.