fix: summarize custom headers
To make space for them, I've removed the process information from jobs.
Note that the element information is still present, so it's easy to find
out what process the job belongs to.

This adds a summary of the custom headers.
- both keys and values are abbreviated to 10 chars
- values containing a key are shortened

Additionally keys prefixed with io.camunda.zeebe are additionally
shortened by dropping the prefix and only taking the first letter of
each part of a camelCase pattern. So camelCase would become 'cc'. This
is useful for the standard user task properties:
- assignee -> a
- candidateGroups -> cG
- candidateUsers -> cU
- dueDate -> dD
- followupDate -> fD
- priority -> p
- userTaskKey -> uTK