feat: remove deleteAllMembers script
This script is not necessary. When a tenant gets deleted the broker exports an individual ENTITY_REMOVED event for each
of the members that are associated with the tenant. We already handle these events by executing a script that removed
the memmbers individually. As a result this deleteAllMembers script will always have 0 results.