fix: store assignee intermediately in separate cf
Storing the assignee in the creating event works, but if we store it
in the user task state directly, then it leads to issues when finalizing
the create command (after the creating listener completed). This flows
through the generalized code of UserTaskProcessor, where we look up both
the current user task, and the intermediate record, and use them to
decide what attributes changed. If we wanted to support it this way, we
needed specialized code for this case (creating/assigning combination)
in this place.

Instead, we can store the intermediate assignee in a separate column
family on the CREATING event, and clean it up on ASSIGNED or
ASSIGNMENT_DENIED.

This ensures that the current user task record and the intermediate user
task record aren't polluted with the assignee, but we can still use the
CREATING event to store it.

A simple look up in the UserTaskCreateProcessor is enough to determine
whether we need to continue assigning.

Note that this means that any corrections to the assignee made by the
creating listener are discarded. We could reject this case separately.
We could consider whether we want to disallow correcting assignee for
all creating listeners, or only those that also have an assigning
listener.