feat: add entity to Tenant using id instead of key
This is part of a refactoring that removes the entityKey property from the TenantRecord.

We need to get rid of all the usages of getEntityKey and setEntityKey. Those need to be replaced with the entityId. For
Mappings this is fairly straightforward as the id already exists. Groups haven't reached this stage of the refactoring
yet. For those we parsing the key to a String and pretending that's the id. This will be refactored later.