feat: more lenient id validation
In old-identity we accept alphanumeric characters, _, - and . characters for tenant ids. Due to migration reasons we
must ensure we accept these same characters in new identity. It isn't sensible to have a different identifier for
tenants specifically, so we will accept these same characters for group, role and mapping ids.

I've decided to also add the @ character. The only difference between usernames and other ids was this character. By
aligning this we don't need a specific pattern anymore.