fix: require default tenant to deploy process
The client must have access to the default tenant in order to deploy a
process using the deprecated deploy process RPC. This RPC can only
deploy to the default tenant, so it is implicitly set. Therefore, the
RPC should be rejected if the client does not have access to the default
tenant.

(cherry picked from commit 9ab1e430a38b5babfaa0249b16835d0c75016f7b)