fix: require default tenant to deploy process
The client must have access to the default tenant in order to deploy a
process using the deprecated deploy process RPC. This RPC can only
deploy to the default tenant, so it is implicitly set. Therefore, the
RPC should be rejected if the client does not have access to the default
tenant.