test: resolve imports and change client id to users
Uncommenting obviously removed all the imports. I had to reimport everything.
I've also changed the client ids to the users that we have now. During this I noticed I missed to add a user earlier,
which is a user that is not assigned to any tenants at all. This has been added as well.