refactor: introduce task listener test helper
This introduces a new TaskListenerTestHelper class, that can store the
helper methods that are currently used extensively in the
TaskListenerTest class. By moving these helper methods into a different
class, we can re-use them more easily.

Note that a lot of the helper methods require access to the EngineRule,
so I've opted to provide this as a constructor argument. To avoid having
to make changes to the methods while moving them, this constructor
argument is used to set the EngineRule as a static field ENGINE. We can
change this to a regular field later.