feat: create job for first `updating` listener on variable update
Enhanced `VariableDocumentUpdateProcessor` to create a job for the first `updating` listener
when a task-scoped variable update is triggered.

Added a test to verify that the job was created and to assert the activated job:
- Custom headers include `"variables"` as a changed attribute.
- Variables include both updated and existing process instance variables.

Introduced `expectPartialUpdate` in `VariableClient` to accommodate the logic split
in `VariableDocumentUpdateProcessor` due to the creation of the job for the first `updating` listener.

Co-authored-by: Ana Vinogradova <ana.vinogradova@camunda.com>