fix: make authorization requests aware of operations on new resources
When an authorization request is made on a new resource
(ex. for a new deployment), the rejection type and error
message may be different as there is no existing resource
to find and evaluate if the user has permissions to
modify. In that case, it is more appropriate to provide
an UNAUTHORIZED rejection type, than a NOT_FOUND. It makes
more sense to move this logic into the AuthorizationCheck
behavior class and the assocaited AuthorizationRequest DTO
than in the processor classes that eventually send back the
rejection.