feat: Increase the default inbound message size to 5MB for clients
We stumbled across an [issue](https://github.com/camunda/camunda/issues/29606)
that made us think raising the limit on the client might be useful
if the payload size approximation on the server is slightly off.
Thus for the client it's still acceptable to process this slightly larger
payload and not raise an error.