test: switch rejections to not found
These tests verify that if a decision gets evaluated with a non-existing decision id or key, we get a rejection. This
rejection checked for an invalid argument. The message indicates no decision was found.
The invalid argument rejection type seems like a bug. We should use the proper rejection type here, which is not found.