feat: improve error message
The forbidden for tenant error message was semantically a bit strange. It would say there were insufficient permissions
to perform an action for a tenant. Permissions here is odd, as tenants aren't about authorizations, but about data
access. I've reworded this message.

Note there is a differences between deploy resource command and deploy command. This is because the deploy command uses
gRPC, whereas the deploy resource command uses rest. In gRPC there is no forbidden status. Hence it is PERMISSION_DENIED
 here.