docs: describe updating a pull request
Now that we'll enforce that pull request updates must not be done using
a merge commit, we should further expand on this in the contributing
guide. This warrants it's own section in the document as it allows us to
easily link to it.

To make this easier for users that are not familiar with using rebase to
achieve this, I've provided a short guide in three steps.

In the description, I've added one additional reason for avoiding merge
commits to update a pull request.

I've also added a note that we encourage contributors to regularly
rebase to minimize the impact of conflicts over time.