docs: explain steps more explicitly
Rather than explaining the why in the workflow run, we can refer to the
contributing guide for this. This means we can be more explicit in the
workflow output, instructing the reader how to overcome the problem.

When they encounter this workflow they have already made the mistake of
updating their pull request with a merge commit, and now need to undo
that.

I want to avoid instructing them git reset --hard $merge_commit^1
because they may have added more commits again after the update.

Instead, we can leave them the note that they need to "remove" the merge
commit. This information is not documented in the CONTRIBUTING guide, as
it's specific to having made the mistake that results in seeing this
workflow output.

We can be explicit about the sha of the merge commit and the base of the
pull request.

If they want to know more information about how to update the pull
request by rebasing, then they can follow the link.