refactor: introduce V2 applier for `UT.CANCELING` to isolate future changes
Creating copies of `UserTaskCancelingApplier` and its test under V2 naming to prepare for upcoming
enhancements related to cleanup of `VariableDocumentState`.

While the cleanup logic could technically be added to the existing applier
(since `MutableVariableState#removeVariableDocumentState` uses `deleteIfExists` in it's implementation),
we will need to introduce a new version anyway to properly support "canceling" user task listeners.

Additionally, it makes no sense to delete `VariableDocumentState` during replay for records created
by previous Zeebe versions, as this data was not persisted before.

Therefore, the original applier is kept as-is (V1), and the cleanup logic will be introduced in the
new V2 version.