refactor: change RoleState to be id-based
Roles are going to be id based. We no longer store the data by key, instead we will use the ids.

For easier refactoring the state class will pretend the key is the id at this time. It will parse key to a string and
use that as id. This will be refactored in later issues.