test: remove the engine from these tests
These tests were accessing the state while the engine was also interacting with the state. This is not thread-safe and
could result in hanging test executions.

It may seem like a lot of changes but it's quite simple. Instead of using the engine and client utility to perform
operations, we now perform them through the appliers directly. I also had to upgrade the tests to Junit5 which is a nice
 bonus.