feat: trigger first `creating` TL on task creation
- if there are no listeners defined, then we go from `creating` to `created` directly
- Added creation of the first `creating` listener job in `UserTaskProcessor` `createUserTask`
- `createUserTask()` now returns the lifecycle state of the user task at the end:
if it returns `CREATED` and the `result` has an `assignee` then we can assign the user task
- unset the `assignee` after `creating` before `created` to ensure we only write it on `creating`.
This will be used in the later tasks for `creating` task listeners as we will store the
`assignee` in `creating` event. To avoid the user task actually being created with an assignee, it should
be cleared in the `created` event.
Please see https://github.com/camunda/camunda/issues/29081 fro more details.