fix: credentials provider tests flakiness
Since the refactor of the credential cache, just removing the cache file after each iteration is not enough anymore.
There is no direct access to the in memory cache map, so using at each tests run a different token doesn't work.
The scope of the tests is not related to the specific value of the token, so having the same string as a token is
correct here