feat: set `assignee` from user task properties in `BpmnUserTaskBehavior`
In order to pass the initial assignee from the user task creation
through `creating` listeners to the `assigning` event, we'll need to
store it for later usage.
As we can only store data in state when we write it in events,
the decision was made to add it to the `creating` event.