test: verify `canceling` task listener job creation
This commit adds a test to verify that a `"canceling"` task listener job is created when a user task is canceled.

Introduces `expectTerminating()` in `ProcessInstanceClient` to await the `PI.ELEMENT_TERMINATING`
event instead of `ELEMENT_TERMINATED`, as termination is now paused when `"canceling"` listeners
are present after creating a job for the first `"canceling"` listener.