feat: remove a limiter on appends based on avg append latency
The condition (now() - (appendLatency.mean() / maxAppendsPerMember) >=
appendTimeStart) may be superfluous as we already limit the number of
inflight requests to maxAppendsPerMember.

Looking into the implementation of DescriptiveStatistics class it's not
been implemented with performance in mind, as it create a lot of
temporary double[] instead of using a fixed size circular buffer.