fix: do not set the commitIndex to be greater than biggest logIndex
It may happen that a follower that is failing to flush entries will
receive a heartbeat from the leaders (an AppendEntries with entries=[]).
Because we don't flush on followers when changing the commitIndex, the
follower will increase its commitIndex.
The next time the leader tries to send an AppendEntriers to the
follower, if it's still unable to flush, it will try to reset the log.
However, because now commitIndex> last logIndex, it will fail to reset
and will become inactive.

(cherry picked from commit 18832069740f77f5dff72a5492178c30dedf97e1)